+++ src/main/java/nextstep/subway/station/domain/Station.java
@@ -1,8 +1,12 @@
 package nextstep.subway.station.domain;
 
-import nextstep.subway.common.BaseEntity;
+import nextstep.subway.global.BaseEntity;
 
-import javax.persistence.*;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
 
 @Entity
 public class Station extends BaseEntity {
